@@ -5,67 +5,74 @@
     "license": "proprietary",
     "minimum-stability": "dev",
     "prefer-stable": true,
+    "repositories": [
+        {
+            "type": "vcs",
+            "url": "https://github.com/jean-gui/symfony-frontend"
+        }
+    ],
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "ext-json": "*",
-        "chrisguitarguy/request-id-bundle": "^4.2",
+        "chrisguitarguy/request-id-bundle": "^v6.0",
         "composer/package-versions-deprecated": "1.11.99.1",
-        "doctrine/annotations": "^1.0",
-        "ekreative/health-check-bundle": "^1.11",
-        "friendsofsymfony/http-cache-bundle": "^2.11",
-        "nelmio/cors-bundle": "^2.1",
-        "nyholm/psr7": "^1.5",
-        "phpdocumentor/reflection-docblock": "^5.2",
-        "sabre/vobject": "^4.3",
-        "sensio/framework-extra-bundle": "^5.1",
-        "strata/symfony-frontend": "^0.8",
-        "symfony/asset": "5.4.*",
-        "symfony/console": "5.4.*",
-        "symfony/dotenv": "5.4.*",
-        "symfony/expression-language": "5.4.*",
-        "symfony/flex": "^1.3.1",
-        "symfony/form": "5.4.*",
-        "symfony/framework-bundle": "5.4.*",
-        "symfony/http-client": "5.4.*",
-        "symfony/intl": "5.4.*",
-        "symfony/mailer": "5.4.*",
-        "symfony/mime": "5.4.*",
-        "symfony/monolog-bundle": "^3.1",
-        "symfony/notifier": "5.4.*",
-        "symfony/process": "5.4.*",
-        "symfony/property-access": "5.4.*",
-        "symfony/property-info": "5.4.*",
-        "symfony/proxy-manager-bridge": "5.4.*",
-        "symfony/security-bundle": "5.4.*",
-        "symfony/security-csrf": "5.4.*",
-        "symfony/serializer": "5.4.*",
-        "symfony/string": "5.4.*",
-        "symfony/translation": "5.4.*",
-        "symfony/twig-bundle": "^5.4",
-        "symfony/validator": "5.4.*",
-        "symfony/web-link": "5.4.*",
-        "symfony/yaml": "5.4.*",
-        "twig/extra-bundle": "^2.12|^3.0",
-        "twig/twig": "^2.12|^3.0",
-        "w3c/website-templates-bundle": "dev-main"
+        "doctrine/annotations": "^2.0",
+        "ekreative/health-check-bundle": "^1.12",
+        "friendsofsymfony/http-cache-bundle": "^3.0",
+        "guzzlehttp/promises": "*",
+        "nelmio/cors-bundle": "^2.5",
+        "nyholm/psr7": "^1.8",
+        "phpdocumentor/reflection-docblock": "^5.4",
+        "sabre/vobject": "^4.5",
+        "strata/symfony-frontend": "dev-main",
+        "symfony/asset": "7.1.*",
+        "symfony/console": "7.1.*",
+        "symfony/dotenv": "7.1.*",
+        "symfony/expression-language": "7.1.*",
+        "symfony/flex": "^2.4",
+        "symfony/form": "7.1.*",
+        "symfony/framework-bundle": "7.1.*",
+        "symfony/http-client": "7.1.*",
+        "symfony/intl": "7.1.*",
+        "symfony/mailer": "7.1.*",
+        "symfony/mime": "7.1.*",
+        "symfony/monolog-bundle": "^v3.10",
+        "symfony/notifier": "7.1.*",
+        "symfony/process": "7.1.*",
+        "symfony/property-access": "7.1.*",
+        "symfony/property-info": "7.1.*",
+        "symfony/proxy-manager-bridge": "6.4",
+        "symfony/security-bundle": "7.1.*",
+        "symfony/security-csrf": "7.1.*",
+        "symfony/serializer": "7.1.*",
+        "symfony/string": "7.1.*",
+        "symfony/translation": "7.1.*",
+        "symfony/twig-bundle": "7.1.*",
+        "symfony/validator": "7.1.*",
+        "symfony/web-link": "7.1.*",
+        "symfony/yaml": "7.1.*",
+        "twig/extra-bundle": "^3.10",
+        "twig/twig": "^3.10",
+        "w3c/website-templates-bundle": "dev-jg/sf7"
     },
     "require-dev": {
+        "overtrue/phplint": "^9.4",
+        "phpunit/phpunit": "^11.2",
+        "rector/rector": "^1.2",
         "roave/security-advisories": "dev-latest",
-        "overtrue/phplint": "^3.0",
-        "phpunit/phpunit": "^9.5",
-        "squizlabs/php_codesniffer": "^3.5",
+        "squizlabs/php_codesniffer": "^3.10",
         "studio24/deployer-recipes": "dev-update-to-deployer-7",
-        "symfony/browser-kit": "^5.4",
-        "symfony/css-selector": "^5.4",
-        "symfony/debug-bundle": "^5.4",
-        "symfony/maker-bundle": "^1.30",
-        "symfony/panther": "^1.0",
-        "symfony/phpunit-bridge": "^5.4",
-        "symfony/stopwatch": "^5.4",
-        "symfony/var-dumper": "^5.4",
-        "symfony/web-profiler-bundle": "^5.4"
+        "symfony/browser-kit": "^7.1",
+        "symfony/css-selector": "^7.1",
+        "symfony/debug-bundle": "^7.1",
+        "symfony/maker-bundle": "^v1.60",
+        "symfony/panther": "^v2.1",
+        "symfony/phpunit-bridge": "^7.1",
+        "symfony/stopwatch": "^7.1",
+        "symfony/var-dumper": "^7.1",
+        "symfony/web-profiler-bundle": "^7.1"
     },
     "config": {
         "optimize-autoloader": true,
