@@ -22,15 +22,13 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * @author Jean-Guilhem Rouel <jean-gui@w3.org>
- *
- * @Route("/news")
  */
+#[Route(path: '/news')]
 class NewsController extends AbstractController
 {
     private const LIMIT = 10;
 
     /**
-     * @Route("/")
      *
      * @param QueryManager        $manager
      * @param Site                $site
@@ -42,6 +40,7 @@ class NewsController extends AbstractController
      * @throws GraphQLQueryException
      * @throws QueryManagerException
      */
+    #[Route(path: '/')]
     public function index(
         QueryManager $manager,
         Site $site,
@@ -91,7 +90,6 @@ class NewsController extends AbstractController
     }
 
     /**
-     * @Route("/{year}/", requirements={"year": "\d\d\d\d"})
      *
      * @param QueryManager        $manager
      * @param int                 $year
@@ -104,6 +102,7 @@ class NewsController extends AbstractController
      * @throws GraphQLQueryException
      * @throws QueryManagerException
      */
+    #[Route(path: '/{year}/', requirements: ['year' => '\d\d\d\d'])]
     public function archive(
         QueryManager $manager,
         int $year,
@@ -159,11 +158,11 @@ class NewsController extends AbstractController
     }
 
     /**
-     * @Route("/{year}/{slug}/", requirements={"year": "\d\d\d\d"})
      *
      * @throws GraphQLQueryException
      * @throws QueryManagerException
      */
+    #[Route(path: '/{year}/{slug}/', requirements: ['year' => '\d\d\d\d'])]
     public function show(
         QueryManager $manager,
         int $year,
