@@ -6,6 +6,7 @@ namespace App\Query\CraftCMS\Ecosystems;
 
 use App\Service\CraftCMS;
 use Strata\Data\Cache\CacheLifetime;
+use Strata\Data\Mapper\MappingStrategyInterface;
 use Strata\Data\Mapper\WildcardMappingStrategy;
 use Strata\Data\Query\GraphQLQuery;
 use Strata\Data\Transform\Data\CallableData;
@@ -36,7 +37,7 @@ class Testimonials extends GraphQLQuery
         $this->site = $site;
     }
 
-    public function getMapping()
+    public function getMapping(): MappingStrategyInterface|array
     {
 
         return [
