@@ -7,6 +7,7 @@ namespace App\Query\CraftCMS\Events;
 use App\Service\CraftCMS;
 use DateTimeImmutable;
 use Strata\Data\Exception\GraphQLQueryException;
+use Strata\Data\Mapper\MappingStrategyInterface;
 use Strata\Data\Query\GraphQLQuery;
 use Strata\Data\Transform\Data\CallableData;
 use Strata\Data\Transform\Value\DateTimeValue;
@@ -72,7 +73,7 @@ class Listing extends GraphQLQuery
         }
     }
 
-    public function getMapping()
+    public function getMapping(): MappingStrategyInterface|array
     {
         return [
             '[id]'               => '[id]',
