@@ -7,6 +7,7 @@ namespace App\Query\CraftCMS\Home;
 use App\Service\CraftCMS;
 use Strata\Data\Exception\GraphQLQueryException;
 use Strata\Data\Mapper\MapArray;
+use Strata\Data\Mapper\MappingStrategyInterface;
 use Strata\Data\Mapper\WildcardMappingStrategy;
 use Strata\Data\Query\GraphQLQuery;
 use Strata\Data\Transform\Data\CallableData;
@@ -44,7 +45,7 @@ class Page extends GraphQLQuery
             ->doNotCache();
     }
 
-    public function getMapping()
+    public function getMapping(): MappingStrategyInterface|array
     {
         $mapping = new WildcardMappingStrategy();
         $mapping->addMapping('heroIllustration', ['[heroIllustration]' => '[heroIllustration][0]']);
