@@ -9,6 +9,7 @@ use DateTimeImmutable;
 use Strata\Data\Cache\CacheLifetime;
 use Strata\Data\Exception\GraphQLQueryException;
 use Strata\Data\Mapper\MapArray;
+use Strata\Data\Mapper\MappingStrategyInterface;
 use Strata\Data\Query\GraphQLQuery;
 use Strata\Data\Transform\Data\CallableData;
 use Strata\Data\Transform\Value\DateTimeValue;
@@ -42,7 +43,7 @@ class RecentActivities extends GraphQLQuery
         ;
     }
 
-    public function getMapping()
+    public function getMapping(): MappingStrategyInterface|array
     {
         return [
             '[category]'         => '[sectionHandle]',
