@@ -7,6 +7,7 @@ namespace App\Query\CraftCMS\PressReleases;
 use App\Service\CraftCMS;
 use Strata\Data\Cache\CacheLifetime;
 use Strata\Data\Exception\GraphQLQueryException;
+use Strata\Data\Mapper\MappingStrategyInterface;
 use Strata\Data\Query\GraphQLQuery;
 use Strata\Data\Transform\Data\CallableData;
 use Symfony\Component\Routing\RouterInterface;
@@ -44,7 +45,7 @@ class Filters extends GraphQLQuery
         ;
     }
 
-    public function getMapping()
+    public function getMapping(): MappingStrategyInterface|array
     {
         return [
             '[archives]' => new CallableData([$this, 'transformArchives'], '[first][year]', '[last][year]')
