@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Query\W3C\Ecosystem;
 
 use App\Service\W3C;
+use Strata\Data\Mapper\MappingStrategyInterface;
 use Strata\Data\Query\Query;
 
 class Groups extends Query
@@ -27,7 +28,7 @@ class Groups extends Query
         return W3C::class;
     }
 
-    public function getMapping(): array
+    public function getMapping(): MappingStrategyInterface|array
     {
         return [
             '[name]'        => '[name]',
