@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Query\W3C\Home;
 
 use App\Service\W3C;
+use Strata\Data\Mapper\MappingStrategyInterface;
 use Strata\Data\Query\Query;
 use Strata\Data\Transform\Data\CallableData;
 
@@ -31,7 +32,7 @@ class Members extends Query
         return W3C::class;
     }
 
-    public function getMapping(): array
+    public function getMapping(): MappingStrategyInterface|array
     {
         return [
             '[name]'        => '[name]',
